feat(.github/workflows/go.yml): add GH action for lint, test, build
Add a GH workflow to lint, test, and build.

The test job runs 'go test' and 'go build'.

The lint job runs golangci-lint and checks that the go.{mod,sum} are up to date.
I added an exclusion rule to avoid changing error messages that exist in cloned
code. Instead of excluding all of format.go, I ran gofumpt on it and fixed a
revive(indent-error-flow) warning.