feat(cmd/wasm): add gh page w/ webassembly celfmt (#12)
Add a new main function that produces a WebAssembly module which can
format programs in the same way as `cmd/celfmt -agent=false`.

The module can be exercised by an accompanying demo page that has a
simple input textarea and 'Format' button. The module exposes build
metadata so that callers can determine the version of elastic/celfmt
that produced the module.

A GitHub action is added to publish the demo page to the project's
github-pages deployment (i.e., https://elastic.github.io/celfmt/). I
chose to keep it quite basic, and not incorporate any third-party
libraries in hopes that this keeps the page very low in maintenance
costs.

Co-authored-by: Dan Kortschak <dan.kortschak@elastic.co>